+++ lib/src/interfaces/interfaces.dart
@@ -8,7 +8,12 @@ import 'package:web3dart/web3dart.dart';
 
 import '../utils/utils.dart' show Uint256;
 import '../web3_signers_base.dart'
-    show P256Credential, PassKeyPair, PassKeySignature, PassKeysOptions;
+    show
+        P256Credential,
+        P256Signature,
+        PassKeyPair,
+        PassKeySignature,
+        PassKeysOptions;
 
 part 'eoa_wallet_interface.dart';
 part 'multi_signer_interface.dart';