+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.6
+
+- Add hardware signer dummy signature
+- Fix Signature conversion method in harware signer
+- Add comparism operators for uint256 class
+
 ## 0.0.5
 
 - Remove unnecessary dependencies